refactor(modules): provide cjs and es6 modules
