+++ README.md
@@ -1,0 +1,15 @@
+[![npm](https://img.shields.io/npm/v/kronos-interceptor-line-parser.svg)](https://www.npmjs.com/package/kronos-interceptor-line-parser)
+[![semantic-release](https://img.shields.io/badge/%20%20%F0%9F%93%A6%F0%9F%9A%80-semantic--release-e10079.svg)](https://github.com/Kronos-Integration/kronos-interceptor-line-parser)
+[![Build Status](https://secure.travis-ci.org/Kronos-Integration/kronos-interceptor-line-parser.png)](http://travis-ci.org/Kronos-Integration/kronos-interceptor-line-parser)
+[![bithound](https://www.bithound.io/github/Kronos-Integration/kronos-interceptor-line-parser/badges/score.svg)](https://www.bithound.io/github/Kronos-Integration/kronos-interceptor-line-parser)
+[![codecov.io](http://codecov.io/github/Kronos-Integration/kronos-interceptor-line-parser/coverage.svg?branch=master)](http://codecov.io/github/Kronos-Integration/kronos-interceptor-line-parser?branch=master)
+[![Coverage Status](https://coveralls.io/repos/Kronos-Integration/kronos-interceptor-line-parser/badge.svg)](https://coveralls.io/r/Kronos-Integration/kronos-interceptor-line-parser)
+[![Code Climate](https://codeclimate.com/github/Kronos-Integration/kronos-interceptor-line-parser/badges/gpa.svg)](https://codeclimate.com/github/Kronos-Integration/kronos-interceptor-line-parser)
+[![Known Vulnerabilities](https://snyk.io/test/github/Kronos-Integration/kronos-interceptor-line-parser/badge.svg)](https://snyk.io/test/github/Kronos-Integration/kronos-interceptor-line-parser)
+[![GitHub Issues](https://img.shields.io/github/issues/Kronos-Integration/kronos-interceptor-line-parser.svg?style=flat-square)](https://github.com/Kronos-Integration/kronos-interceptor-line-parser/issues)
+[![Stories in Ready](https://badge.waffle.io/Kronos-Integration/kronos-interceptor-line-parser.svg?label=ready&title=Ready)](http://waffle.io/Kronos-Integration/kronos-interceptor-line-parser)
+[![Dependency Status](https://david-dm.org/Kronos-Integration/kronos-interceptor-line-parser.svg)](https://david-dm.org/Kronos-Integration/kronos-interceptor-line-parser)
+[![devDependency Status](https://david-dm.org/Kronos-Integration/kronos-interceptor-line-parser/dev-status.svg)](https://david-dm.org/Kronos-Integration/kronos-interceptor-line-parser#info=devDependencies)
+[![docs](http://inch-ci.org/github/Kronos-Integration/kronos-interceptor-line-parser.svg?branch=master)](http://inch-ci.org/github/Kronos-Integration/kronos-interceptor-line-parser)
+[![downloads](http://img.shields.io/npm/dm/kronos-interceptor-line-parser.svg?style=flat-square)](https://npmjs.org/package/kronos-interceptor-line-parser)
+[![Commitizen friendly](https://img.shields.io/badge/commitizen-friendly-brightgreen.svg)](http://commitizen.github.io/cz-cli/)
