+++ package.json
@@ -2,13 +2,14 @@
   "name": "kronos-interceptor-line-parser",
   "version": "0.0.0-semantic-release",
   "description": "Splits a stream into lines.",
-  "main": "index.js",
+  "main": "dist/interceptor.js",
+  "module": "src/interceptor.js",
   "scripts": {
     "cover": "npm run build && node_modules/.bin/istanbul cover node_modules/.bin/_mocha -- -u exports tests/*_test.js",
     "doc": "./node_modules/.bin/jsdoc lib/*.js",
     "test": "npm run build && node_modules/.bin/mocha tests/*_test.js && markdown-doctest",
     "semantic-release": "semantic-release pre && npm publish && semantic-release post",
-    "build": "node_modules/rollup/bin/rollup --output=index.js -c -- {{module}}",
+    "build": "node_modules/rollup/bin/rollup --output=dist/interceptor.js -c -- src/interceptor.js",
     "docs": "jsdoc2md -l off -t doc/README.hbs -f src/*.js > README.md"
   },
   "repository": {
@@ -26,7 +27,6 @@
     "chai": "^3.5.0",
     "cz-conventional-changelog": "^1.2.0",
     "istanbul": "^0.4.5",
-    "jsdoc": "3.4.2",
     "kronos-test-interceptor": "^2.0.5",
     "mocha": "^3.2.0",
     "semantic-release": "^6.3.5",
@@ -34,16 +34,13 @@
     "markdown-doctest": "^0.9.1",
     "rollup": "^0.38.0"
   },
-  "contributors": [
-    {
-      "name": "Torsten Link",
-      "email": "torstenlink@gmx.de"
-    },
-    {
-      "name": "Markus Felten",
-      "email": "markus.felten@gmx.de"
-    }
-  ],
+  "contributors": [{
+    "name": "Torsten Link",
+    "email": "torstenlink@gmx.de"
+  }, {
+    "name": "Markus Felten",
+    "email": "markus.felten@gmx.de"
+  }],
   "license": "BSD-2-Clause",
   "engines": {
     "node": ">=6.9.2"
